refactor(daemon): group app-config.ts + installation.ts into config/
