+++ apps/daemon/src/config/installation.ts
