+++ apps/daemon/src/connectionTest.ts
@@ -40,7 +40,7 @@ import { createClaudeStreamHandler } from './runtimes/claude-stream.js';
 import { diagnoseClaudeCliFailure } from './claude-diagnostics.js';
 import { createCopilotStreamHandler } from './copilot-stream.js';
 import { createJsonEventStreamHandler } from './runtimes/json-event-stream.js';
-import { agentCliEnvForAgent, validateAgentCliEnv } from './app-config.js';
+import { agentCliEnvForAgent, validateAgentCliEnv } from './config/index.js';
 import {
   antigravityAuthGuidance,
   antigravityQuotaGuidance,
@@ -57,7 +57,7 @@ import {
   isUnsupportedMaxTokensError,
 } from './integrations/openai-chat-token-params.js';
 import { aihubmixHeaders } from './integrations/aihubmix.js';
-import type { AgentCliEnvPrefs } from './app-config.js';
+import type { AgentCliEnvPrefs } from './config/index.js';
 import type { RuntimeAgentDef } from './runtimes/types.js';
 import { preparePromptFileForAgent, type PreparedPromptFile } from './runtimes/prompt-file.js';
 import { configuredAllowedInternalHosts } from './origin-validation.js';
