+++ apps/daemon/src/orbit.ts
@@ -4,7 +4,7 @@ import path from 'node:path';
 
 import type { OrbitRunSummary, OrbitStatusResponse } from '@open-design/contracts/api/orbit';
 
-import type { OrbitConfigPrefs } from './app-config.js';
+import type { OrbitConfigPrefs } from './config/index.js';
 import { skillCwdAliasSegment } from './cwd-aliases.js';
 
 export interface OrbitConnectorRunResult {
